std: Allow use of functions w/ parameters in std.Once
Given the lack of closures the only way to supply some extra arguments
to a function wrapped in Once is to use global variables. While this
works in theory, we can do better than hoping for the user to protect
the globals from being concurrently modified by allowing the function to
have zero or more arguments.