FixedSlice: a simple way to represent small data whose max size is known
This is a simple structure containing an array and a length, that can
be viewed as a slice.

It is useful to pass-by-copy small data whose exact size is known at
runtime, but whose maximum size is known at comptime. This greatly
simplifies code that otherwise would require an allocator, or
reimplementing what this type does.